+++ tests/cli/test_session_commands.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
 from starlite import Starlite
@@ -13,18 +15,10 @@ if TYPE_CHECKING:
     from click.testing import CliRunner
     from pytest_mock import MockerFixture
 
-    from starlite.storage.memory import MemoryStorage
 
-
-def test_get_session_backend(memory_storage_backend: "MemoryStorage") -> None:
-    session_middleware = ServerSideSessionConfig(storage=memory_storage_backend).middleware
-    app = Starlite(
-        [],
-        middleware=[
-            RateLimitConfig(rate_limit=("second", 1)).middleware,
-            session_middleware,
-        ],
-    )
+def test_get_session_backend() -> None:
+    session_middleware = ServerSideSessionConfig().middleware
+    app = Starlite([], middleware=[RateLimitConfig(rate_limit=("second", 1)).middleware, session_middleware])
 
     assert get_session_backend(app) is session_middleware.kwargs["backend"]
 
@@ -49,8 +43,8 @@ def test_delete_session_cookie_backend(runner: "CliRunner", monkeypatch: "Monkey
 def test_delete_session(
     runner: "CliRunner", monkeypatch: "MonkeyPatch", mocker: "MockerFixture", mock_confirm_ask: "MagicMock"
 ) -> None:
-    monkeypatch.setenv("STARLITE_APP", "docs.examples.middleware.session.memory_storage:app")
-    mock_delete = mocker.patch("starlite.storage.memory.MemoryStorage.delete")
+    monkeypatch.setenv("STARLITE_APP", "docs.examples.middleware.session.file_store:app")
+    mock_delete = mocker.patch("starlite.stores.file.FileStore.delete")
 
     result = runner.invoke(cli_command, ["sessions", "delete", "foo"])
 
@@ -79,8 +73,8 @@ def test_clear_sessions_cookie_backend(runner: "CliRunner", monkeypatch: "Monkey
 def test_clear_sessions(
     runner: "CliRunner", monkeypatch: "MonkeyPatch", mocker: "MockerFixture", mock_confirm_ask: "MagicMock"
 ) -> None:
-    monkeypatch.setenv("STARLITE_APP", "docs.examples.middleware.session.memory_storage:app")
-    mock_delete = mocker.patch("starlite.storage.memory.MemoryStorage.delete_all")
+    monkeypatch.setenv("STARLITE_APP", "docs.examples.middleware.session.file_store:app")
+    mock_delete = mocker.patch("starlite.stores.file.FileStore.delete_all")
 
     result = runner.invoke(cli_command, ["sessions", "clear"])